feat: Add pydantic validation
