+++ setup.py
@@ -7,6 +7,6 @@ setup(
     author="Aleksander Fidelus",
     author_email="alek.fidelus@gmail.com",
     packages=find_packages(),
-    install_requires=["github-action-utils==1.1.0"],
+    install_requires=["github-action-utils==1.1.0", "pydantic==2.0.3"],
     extras_require={"all": DEV + TEST, "dev": DEV, "test": TEST},
 )
