+++ tests/test_model.py
@@ -1,11 +1,11 @@
-from validator.main import Action
+from validator.model import ActionInput
 import pytest
 
 
 @pytest.mark.parametrize("regex_match_type", ["not_supported_match_type", "MATCH"])
 def test_validate_incorrect_input(regex_match_type: str):
     with pytest.raises(SystemExit) as exc_info:
-        Action(
+        ActionInput(
             regex_match_type=regex_match_type,
             regex_pattern="^test",
             text="test_validate_input",
